+++ modules/portal-security/portal-security-sso-opensso/src/main/java/com/liferay/portal/security/sso/opensso/configuration/definition/OpenSSOCompanyServiceConfigurationPidMapping.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.opensso.module.configuration.definition;
+package com.liferay.portal.security.sso.opensso.configuration.definition;
 
-import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
-import com.liferay.portal.security.sso.opensso.module.configuration.OpenSSOConfiguration;
+import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
+import com.liferay.portal.security.sso.opensso.configuration.OpenSSOConfiguration;
+import com.liferay.portal.security.sso.opensso.constants.OpenSSOConstants;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -23,12 +24,17 @@ import org.osgi.service.component.annotations.Component;
  * @author Mika Koivisto
  */
 @Component
-public class OpenSSOCompanyServiceConfigurationBeanDeclaration
-	implements ConfigurationBeanDeclaration {
+public class OpenSSOCompanyServiceConfigurationPidMapping
+	implements ConfigurationPidMapping {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
 		return OpenSSOConfiguration.class;
+	}
+
+	@Override
+	public String getConfigurationPid() {
+		return OpenSSOConstants.SERVICE_NAME;
 	}
 
 }
