+++ modules/portal-security/portal-security-sso-opensso/src/main/java/com/liferay/portal/security/sso/opensso/internal/OpenSSOImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.opensso;
+package com.liferay.portal.security.sso.opensso.internal;
 
 import com.liferay.portal.kernel.io.unsync.UnsyncBufferedReader;
 import com.liferay.portal.kernel.log.Log;
