+++ modules/portal-security/portal-security-sso-opensso/src/main/java/com/liferay/portal/security/sso/opensso/internal/portlet/action/PortalSettingsOpenSSOFormMVCActionCommand.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.opensso.portlet.action;
+package com.liferay.portal.security.sso.opensso.internal.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.security.sso.opensso.constants.OpenSSOConstants;
