+++ modules/portal-security/portal-security-sso-opensso/src/main/java/com/liferay/portal/security/sso/opensso/internal/servlet/taglib/PortalSettingsOpenSSOAuthenticationDynamicInclude.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.opensso.servlet.taglib;
+package com.liferay.portal.security.sso.opensso.internal.servlet.taglib;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
