+++ modules/portal-security/portal-security-sso-opensso/src/main/resources/META-INF/resources/com.liferay.portal.settings.web/init.jsp
@@ -30,8 +30,8 @@ page import="com.liferay.portal.kernel.settings.ParameterMapSettingsLocator" %><
 page import="com.liferay.portal.kernel.util.ParamUtil" %><%@
 page import="com.liferay.portal.kernel.util.Validator" %><%@
 page import="com.liferay.portal.security.sso.OpenSSOUtil" %><%@
-page import="com.liferay.portal.security.sso.opensso.constants.OpenSSOConstants" %><%@
-page import="com.liferay.portal.security.sso.opensso.module.configuration.OpenSSOConfiguration" %>
+page import="com.liferay.portal.security.sso.opensso.configuration.OpenSSOConfiguration" %><%@
+page import="com.liferay.portal.security.sso.opensso.constants.OpenSSOConstants" %>
 
 <%@ page import="java.util.ArrayList" %><%@
 page import="java.util.List" %>
